refactor(text-input): rename TextInput directory
