+++ components/index.ts
@@ -3,6 +3,6 @@ export * from "./button";
 export * from "./card";
 export * from "./code";
 export * from "./heading";
-export * from "./input";
 export * from "./provider";
 export * from "./text";
+export * from "./text-input";
